+++ lib/src/route_management/w_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
+import 'package:nested/nested.dart';
 
 /// Model class for named route.
 class WPage {
@@ -9,7 +10,10 @@ class WPage {
   final String? title;
 
   /// Register service.
-  final void Function(BuildContext context, dynamic arguments)? serviceBuilder;
+  final List<SingleChildWidget> Function(
+    BuildContext context,
+    dynamic arguments,
+  )? providerBuilder;
 
   /// Builds the primary contents of the route.
   final Widget Function(BuildContext context, dynamic arguments) builder;
@@ -17,7 +21,7 @@ class WPage {
   WPage({
     required this.path,
     this.title,
-    this.serviceBuilder,
+    this.providerBuilder,
     required this.builder,
   });
 }