+++ lib/src/service_locator/index.dart
@@ -1,3 +1,0 @@
-export './w_disposable.dart';
-export './w_service_builder.dart';
-export 'w_service/w_service.dart';