+++ lib/src/state_management/index.dart
@@ -1,5 +1,3 @@
-export './extensions/value_notifier_extensions.dart';
-export './extensions/value_notifiers_extensions.dart';
 export './w_listener.dart';
 export './w_multi_listener.dart';
 export 'extensions/list_extensions.dart';