+++ lib/src/w_disposable.dart
@@ -1,0 +1,3 @@
+abstract class WDisposable {
+  void dispose() {}
+}