+++ lib/witt.dart
@@ -1,5 +1,8 @@
 library witt;
 
 export 'src/route_management/index.dart';
-export 'src/service_locator/index.dart';
 export 'src/state_management/index.dart';
+export 'src/w_disposable.dart';
+export 'src/w_multi_provider.dart';
+export 'src/w_provider.dart';
+export 'src/w_provider_base.dart';